// src/app/models/search-item.ts
export interface SearchItem {
    country: string;
    lng: number;
    city: string;
    searchType: string;
    alternative: string[];
    index: number;
    bookingId: string;
    placeType: string;
    placeKey: string;
    iata: string;
    countryIso: string;
    locationId: string;
    name: string;
    ufi: number;
    isPopular: boolean;
    region: string;
    lang: string;
    lat: number;
}